+++ src/components/ShoppingCartPage/CartItem/CartItem.style.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import theme from '../../styles/theme';
+import theme from '../../../styles/theme';
 
 export const Layout = styled.div`
   display: flex;
