+++ src/components/ShoppingCartPage/PriceContainer/PriceContainer.style.tsx
