+++ src/components/ShoppingCartPage/TotalPriceContainer/TotalPriceContainer.style.tsx
