+++ src/components/ShoppingCartPage/TotalPriceContainer/TotalPriceContainer.tsx
