+++ src/pages/ShoppingCartPage/ShoppingCartPage.tsx
@@ -2,10 +2,10 @@ import { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import { useLoaderData, Await, Link } from 'react-router-dom';
 import Header from '../../components/Header/Header';
-import TitleContainer from '../../components/Container/TitleContainer/TitleContainer';
-import SubmitButton from '../../components/Button/SubmitButton/SubmitButton';
-import CartItemList from '../../components/CartItemList/CartItemList';
-import TotalPriceContainer from '../../components/Container/TotalPriceContainer/TotalPriceContainer';
+import TitleContainer from '../../components/common/TitleContainer/TitleContainer';
+import SubmitButton from '../../components/common/SubmitButton/SubmitButton';
+import CartItemList from '../../components/ShoppingCartPage/CartItemList/CartItemList';
+import TotalPriceContainer from '../../components/ShoppingCartPage/TotalPriceContainer/TotalPriceContainer';
 import type { TCartItem } from '../../types/CartItem.type';
 import { selectedCartItemListState } from '../../recoil/atoms/atoms';
 import { fetchCartItemList } from '../../apis';
